updating helpers from the PR suggestions
